new file:   ft_lstlast.c

+++ LIBBEFFETTI/ft_bzero.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned long n)
+void	ft_bzero(void *s, size_t n)
 {
 	char			*p;
 	unsigned int	i;
